Cast to signed before float conversion.
Data sheet is ambiguous on this, though I suppose signed does make
more sense. Thanks to Pete Marchetto for finding this.

+++ modules/water/main/oem_atlas.cpp
@@ -219,7 +219,7 @@ bool OemAtlas::read(float *values, size_t &number_of_values) {
     }
 
     for (auto i = 0; i < cfg.number_of_values; ++i) {
-        float value = __builtin_bswap32(raw[i].u32);
+        float value = (int32_t)__builtin_bswap32(raw[i].u32);
         value /= cfg.divisor;
         values[i] = value;
     }
